Backend/requirements_dev.txt
django
django-cors-headers
djangorestframework
djangorestframework-jwt
django-extra-fields
mysqlclient
mysql-connector
cmake
dlib
opencv-python
face_recognition
numpy
Pillow
scipy
mtcnn
tensorflow
